Added basic functionality for three bosses.
The basic functionality of each boss is in place, but they can only move like normal zombies and damage by touch like other zombies.

Corrected inBounds method of TransactionButton because it was offset by the size divided by two, causing the bounding box to be off to the right and below where it should be.

Corrected font and positioning of menu buttons on Game Over screen.

Added console spawning commands for the three bosses.