Fixed Reloading Bug
If the player's ammo in the current clip and the ammo in their inventory were both smaller than the clip size, you could press reload repeatedly and it would just swap the two values.

Also modified color and draw position of Zombat blood stream so it appears to come from its mouth and the red isn't so bright.