Added "Buy Ammo" Button
Player can now buy ammo for individual weapons for a set price via the shop. There is a button below the weapon description when a weapon in the player's inventory is selected that will allow the player to buy a clip of ammo at a time for a set price.