Bug Fixes, Boss Waves, and Enemy Wave Balancing
Mostly just bug fixes, but also implemented boss waves so that no other enemies spawn with the boss (this is subject to change), and balanced a few things with enemies and waves in general.

Did some playtesting and found some new bugs to fix and things to balance.

Refactored some code.