Balancing, Bug Fixes, and Ammo Clip Max
Put a max cap on the number of clips the player can have for each weapon. Ammo buy button will be grayed out if the player has the max amount of ammo.

Added bleed effect to Stitches' hook so it's more dangerous and gives incentive for player to get close to him so he'll let go.

Added console commands to control music playback.

Spawn rate will now progressively get faster, causing enemies to spawn on the map closer together in time.

Increased Zombat health and number of Zombats in a boss wave.