Spawning Method Update & Flashlight
Method of randomly determining enemies to spawn has been updated so methods that determine if an enemy CAN BE spawned do not use string comparisons, which should be faster. It's also cleaner.

Playe rnow has a spotlight on them, but the flashlight in front of them still does not work because OpenGL blending is silly.