Bosses Immune to Knockback, First Easter Egg, and More
Knockback effect is broken as hell during a boss fight, as the Assault Rifle is strong enough to keep them from getting anywhere near you.

Decreased the semi-auto fire cooldown of the Pistol by half so it can be fired as fast as the user can click the button (almost).

Increased damage done by Gasbag's poison cloud because it was basically harmless.

Added the game's first easter egg. :)