UI Changes & Bug Fixes
Shop now displays each weapon's current ammo count beneath its portrait in the player's inventory. There is now a button to buy the max amount of ammo you can.

The HUD now includes a display of the player's current cash.

Fixed (hopefully) a bug in which continuing to shoot at the location of a dropped powerup after an enemy dies would cause the powerup to get knocked back as if you were shooting it.

Increased default flashlight size.

Music now defaults to half volume; however, when the settings menu is implemented, the player should be able to manually change the music volume.

Fixed a bug where vanishing messages would not be cleared in between games.