Burning Particle Effect
Added visual effect for burning status so fire particles randomly fly off of the body of the entity that is burning. No image for particles yet, so they're just orange blobs. Need to add an image for them. Burning effect makes use of custom Emitter class.

Various refactored code changes to support this new effect, including a method for the StatusProjectile class that is called when the projectile hits the Entity, allowing the projectile access to the entity that it hit.

Also added emission interval to emitters because I forgot to do that in the initial write-up of the class. Particles are now automatically emitted in the update method according to the emission interval. There is also a method that can be overridden that allows the starting position of the emitted particle to be transformed.

Status effect now has access to the entity it is being applied to in the handleEntity() method.