Bug Fixes, Loot Tables & Settings
Made some minor tweaks to loot table values so that item drops aren't as ridiculous. Time will tell if these values are suitable.

Fixed a minor bug with console input that was causing the user to only be able to type the keys that are mapped in the controls layout.

Knockback should now occur in the direction that the projectile that hit the enemy was traveling in.

Added several settings screens, only one of which (audio settings) has a functional screen. Added "Slider" class for sliding property values. They work insofar as you can click them to change their value, but the drag functionality does not work.