Sound & Display Settings
Added new volume slider to audio settings for sound effect volume.

Added custom "shadow opacity" settings screen for determining how dark the shadow layer that the flashlight cuts through should be. There is a minimum and maximum darkness value to avoid the player choosing values that look strange (an opacity of 1.0 is completely pitch black and the flashlight is blue for some reason). The screen shows a basic layout with the player, flashlight, and some zombies in a field so you can see the effect in real-time. Need to add text to explain what this setting does though.