Floating Damage Text, Criticals, Burning & More
The damage done by the player's weapons will now shoot out of the enemies taking damage.

Critical hits have been added. They have a 5% chance to do double damage. Eventually there will be powerups to increase critical chance and/or damage and maybe perks.

The burning status effect no longer does damage every update, and instead does damage every second equal to the amount of damage done in that amount of time according to how much damage it was doing per update before.

Realized the wrong key was assigned to the MOVE_UP AND MOVE_LEFT controls and swapped them. There are also other issues with the Controls class -- mostly just that certain characters (such as space) display the typed character instead of the name of the key (ie: space shows blank in the Controls screen when assigned instead of the word "space").