Damage Balancing, Variable Particle Colors & Menu Icons
I've updated all weapons to try to balance their damage ranges so they're not over/under powered. The flamethrower has been changed so that individual particles no longer do damage, and the damage comes from the burning status effect. This needs to be fine-tuned.

Particles can now have a "color generator", which can generate random colors between two starting colors rather than making all particles from an emitter be the same color.

Added two icons to the bottom-right of the gameplay screen to indicate what buttons to press to access the shop and training screen menus, along with text next to them with the key currently mapped to those menus. The player's money display has been moved to the bottom-center of the screen.

Other minor tweaks, such as reducing the number of Zombats in a wave, as they're ridiculously deadly.