Armor, Flashlight Toggle, Shop Fix & More
Added armor item that adds either 50 or 100 armor depending on the type of armor. This will need to be expanded upon, as it is currently not accessible anywhere except the console.

Fixed the buy ammo buttons in the shop, as the max ammo button would refill the current clip, allowing you to expend your clip, then max your ammo to avoid reloading. Also renamed the "buy ammo" button to "One Clip" to indicate that you're purchasing one clip. Also modified it so that if your ammo inventory is full and you have less than a clip in the current magazine, you'll only be charged for the price of the missing ammo as opposed to a full clip.

Updated powerup drop rates.

Added translation function to translate special characters that do not have an obvious character (ie: space, shift, tab, etc.

The flashlight can now be toggled by pressing F (by default).

The player's weapon will now no longer automatically reload upon expending a clip.