Screen Shake, Armor in Shop & Level-Up Sound
The screen now shakes a variable amount while firing certain weapons and while player-generated explosions go off. Added a new Camera class to implement this, which can be used in the future for removing the restriction of the limited play area.

Armor can now be bought from the shop. Currently, only regular armor (value of 50) is available for $1,000.

A lightning / thunderclap sound now plays when the player levels up.