Particle Enhancements, Fading HUD & Refactoring
Enhanced the particle class to enable particles to have animations. First example of this is an animated flame particle for the Flamethrower and burning status effect. The visual isn't very good, but it's good enough for now.

Certain elements of the HUD will now fade if the player is behind them so that the player does not lose track of where they are.

Moved certain elements of the HUD to their own classes to condense the HUD class.

Darkness is now less dark.

Poison damage now ignores the player's armor.