Rendering Layers, Damage Vignette & Tweaks
Game entities now have layers assigned to them to determine the order that they are rendered in. This is to correct issues with the non-guaranteed order of the map the entities are stored in where some entities were being incorrectly drawn over others (ie: blood particles being drawn on top of the enemies and player). This could also help solve some issues in the future.

Added a "damage vignette" to the camera so that when the player takes damage, there is a short transparent red overlay to indicate that the player is taking damage. This along with the subtle screen shake and grunt noises should be more than enough to warn the player.

Decreased size of particles in the BURST blood generator so that blood splatters are less ridiculous.

Halved the Bow & Arrow's charge time, as it's basically useless as it is. Will need to tweak further, and possibly apply speed multiplier to charge rate?

Enemies now draw particles before themselves so that particles are not drawn on top of them.

Added a couple helper functions to the Controls class that can be used if I decide to migrate to polled input, which could solve some issues with the game not registering input properly during fade-in transitions, requiring the user to release the key and press it again to register the input.