Config File & Armor Availability
The game's settings can now be stored in a config file in settings/.config. This has yet to be tested in the scope of the runnable JAR, however, so revisions may be made.

Armor is now only able to be purchased once from the shop, after which is disappears. After each wave, there is a 20% chance that more armor of a random type will become available in the shop.

Corrected a dumb little graphical hiccup where the color of text wasn't being set before drawing the game version on menus, so if you hovered over the back button, the version would become the same color as the back button.