Focus Loss & Other Bug Fixes
The main update loop will now check to see if the window has lost focus, and if so, will not update.

Was able to recreate the issue of enemies not spawning / the enemy controller not properly transitioning to the break period, but not able to pinpoint the reason. Added new console command "/ec" which will print out the number of unborn, alive, and "add immediately" enemies to possibly diagnose the problem.

Fixed the Big Red Button's update method, as it was not able to reload.

Fixed issue where player's flashlight layer wasn't being drawn during respawn period (was only being drawn if player was alive).