Right Mouse Melee & Refactoring
The melee weapons have been separated from ranged weapons and will not show in the scroll wheel as part of your loadout. Instead, the equipped melee weapon is bound to right mouse click and can be used alongside your currently equipped ranged weapon.

Refactored EnemyController class to be a singleton so it's accessible anywhere. Might as well make everything a singleton at this point...

Corrected draw order of boss health bar so it's now above the shadow mask and not darkened.