Molotov & Refactoring
The Molotov is now available in the shop. It will create a small patch of fire to apply the burning status to enemies that enter the area.

Had to change the way ranged weapons are fired so charged weapons are ignored. Instead, charged weapons handle their own "firing" mechanism in their update function. This fixes the problem of charged weapons continuously reporting "out of ammo" after firing the last shot even when the mouse is not being held down.

Also fixed refactoring error from before where the player's update method was retrieving the current melee weapon always as the current weapon.