New Weapons, Projectile Penetration & Bug Fix
Added the Bastard Sword, a multi-hit sword with long reach. Also added the Crossbow and Remington hunting rifle. 

Added projectile penetration, which has been given to the Remington so that it can penetrate one enemy.

Fixed a mix-up in logic where I was applying the damage bonus from the damage skill to ranged weapons twice.