Revolver & Stinger, Animation Fix
Added the Revolver. Just a more powerful pistol, really.

Added the Stinger, a rocket launcher with big booms.

While working on Stinger, noticed that when a particle has an animation set, it rotates the image twice... once in the Particle.render() method, and once in the Animation.render() method. Changed Particle.render() so that it only rotates if a plain image is being used. Now all particles with an animation by default subtract 90 degrees from their rotation.