Enemy Attacks & Gag Weapon
Enemies now attack periodically as opposed to doing constant damage while touching the player. Each enemy has their own damage cooldown and uses the Dice class to roll damage for each attack.

Added to the Zumby sprite sheet to include a basic attack animation for the new attack behavior. Will need to add this for all other melee attackers.

Added the Lollipop, a gag weapon, which does a lot of damage, but also causes confetti to shoot out of the enemy and make a stupid birthday noise maker sound effect.