Level-Restricted Weapons & Rotdog Animation
Weapons now have level restrictions and will only appear in the shop once the player meets that level requirement. There is also a degree of randomness as to whether or not each weapon will appear in the shop.

The Rotdog now has a (terrible) attack animation like the Zumby.

The text displayed in the top-right of the HUD in between waves has been changed to be more informative.