New Enemy: Starfright
Added new enemy type: Starbright. This enemy will charge at you at high speed, but is photosensitive and moves slower when the player shines their flashlight on it. If it gets close enough, it will explode like a flashbang, deafening and damaging the player. If the player is looking at Starbright when this happens, they will also be temporarily blinded. The deafening effect leaves a ringing in the player's ears, drowning out other sounds and the game music.