New Enemy Type: El Salvo
Added new enemy type: El Salvo. He will charge the player and quickly explode when within range. The explosion is highly damaging and is enough to kill a player without upgraded health. The damage is reduced if the player is far enough away from the center of the blast, though.

Fixed scaling of explosion entities both in render method and in creation of collider.