Status Immunities
Status handler class now supports status immunities. For now, this really only applies to bosses, but as more status effects are added, this will be used more. All bosses are currently immune to paralysis and poison.

Made small correction to Status Projectile class, which was calling status effect onApply method, despite the fact that adding a status also does this.

Added "noEffect" method to status effect class to define behavior when a status has no effect on the entity it has attempted to apply itself to (ie: immunities).