Prowler Enemy Type, Ammo Warning Messages, Acid Status & Tweaks
Added new enemy type: The Prowler. This enemy will crawl towards you from the darkness, and once within range, lunge at you. If he hits you, you'll be pinned/paralyzed for several seconds while he, and any other zombies nearby, pummel you.

Corrected ammo warning texts so that the proper message is shown when you are completely out of ammo, need to reload, or are currently reloading, as opposed to always saying "Out of Ammo".

Added Acid status effect so that Upchuck and Aberration now apply a DoT effect instead of doing direct damage per projectile. This is to keep everything in line with the new d20-based damage system.

A message will now appear when a new weapon is added to the shop to inform the player.