Stamina & Achievements
The player now has a Stamina bar showing how much stamina the player has to use for melee attacks. There is a flat refresh rate of 10% per second for now, but this will be upgradeable.

Added Achievements. There are only a few basic kill achievements for now, but the notifications will show up at the bottom-right above the shop and training icons when the player fulfills an achievement. Achievements are not yet saved, but that will be implemented shortly.

Fixed the way the Calculate.TextWrap() method words so that it does not overflow the max width.