State-Based Achievements
Added state-based achievements, which allow an achievement to keep track of its progress towards completion via a finite state machine of sorts. The transitions between states can either match a posted metric or evaluate a lambda function against the posted metric to allow for complicated conditions.

Refactored achievement controller to make posted metrics long values so that the values can be composed together using bitwise OR to allow for complicated metrics and speed up comparisons.

Added pre-defined metrics which are powers of 2. They need to be powers of 2 to be compose-able, otherwise you could have collisions depending on which values you compose.

Added metrics to various game events throughout the code.

Added a few achievements.

Added a new crosshair.