Achievement Display Menu & Updated Metrics
Added new menu for displaying player's achievement progress. Still need actual icons for each achievement and to display some sort of indicator either that an achievement is complete or not complete, or even hidden.

Revised the metrics available to generalize more and allow for more combinations of metrics.

Added new global so states can make a distinction of whether they came to the current state from an active game or not. This is so menus can be made accessible from the ESC menu in-game.

Reduced the size of the player's collider.