Talents Screen UI Changes
The current and next player level are now displayed to the left and right of the progress bar, respectively.

There are now tooltips for the talents. However, because tooltips currently only support one body of text, they are unable to display both the talent name and the description. Need to modify tooltips to have a header and body.