Talents Screen Functionality
Talents screen now displays the number of skill points the player has and has accept and exit buttons. The accept button will accept any changes to the player's talents and lock them in, while the exit button will not only exit the screen, but discard any talent changes the player has made.

All that's left is to add images for the talents and provide functionality for each talent throughout the game code.

Fixed issue with soundtrack restarting when returning to main menu from anywhere.