Talents Screen Enhancements
The tooltips for talent buttons now have the name of the associated talent as well. They will also display the number of ranks the player has next to the max number of ranks for that talent below the button.

Increased spacing between rows of talents to accommodate talent progress text.

Tooltips will now auto-adjust to the height of the text inside them instead of using a constant height.