Added Stasis Talent
Added the Stasis talent for the Tactics tree. Refactored the Enemy class to allow for setting and resetting enemy speeds to allow this talent to manipulate enemy speeds.

Added the Slow status effect, which is applied to all enemies upon picking up the Speed-Up powerup when you have the Stasis talent, simulating time "slowing down".

Changed icon for Speed-Up to be green so Slow Down icon could be red.