Talents and Bug Fixes
Implemented the Stockpile and Sustainability talents. Some minor refactoring went into this.

Corrected a typo in the Relentless talent that was giving an incorrect amount of stamina back when the talent triggered.

Corrected incorrect behavior on Talents screen causing the state to always dispose of talent changes upon clicking the exit button whether or not any changes had been made. Also added this discard logic to the Talents screen shortcut key.