Talents and Achievements
Added the Firepower and Durability talents. These will need an implementation for future constructs.

Refactored state-based achievements to make them more robust and allow for waiting for multiple instances of a metric before transitioning. Also added support for "branch" state transitions in which a transition is always made on a state depending on the condition or presence of the given metric.

Added the "It's Time To Party!" achievement.