Bug Fixes and UI Enhancements
Corrected an issue with the EnemyController (well, less of an issue and more that my dumb ass forgot to add Starfright, El Salvo, and the Prowler to the spawn list).

Made correction to Baseball Bat class because its price wasn't returning the correct value.

Adjusted some UI elements on the Talents and Shop screens and added some new elements to the Talents screen to make it more informative.

Made it impossible to sell the Machete and Beretta so you can't sell all your weapons and end up with nothing.