Crafting Changes, New Recipe, Reformatting
The Molotov has been removed as a crafting recipe and will instead be made available in the shop again. Instead, a recipe for the Spiked Bat has been added for testing purposes until the rest of the crafts are finished.

Reformatted the crafting architecture so that instead of weapon costs being represented as strings with the weapon's name, an enum containing the name, description, and image name of that weapon is used so that the craft windows can retrieve that information. This might also make the inventory system more flexible later on.