Pipe Bomb, Molotov & Crafting Window
Added the Pipe Bomb basic craft recipe.

Modified Recipe class to allow for "repeatable" crafts, so that if the recipe is repeatable and you already have the weapon in your inventory, it will instead add a clip of ammo to the weapon. Because of this, Molotov is now a crafted weapon again instead of a weapon from the shop.

Corrected the crafting display so that you cannot scroll infinitely upwards. Also, the list of crafts is refreshed when you craft something so you don't have gaps in the list.