New Images & Corrections
There is a new arm position image to go along with the weapon image for the MP5 as well as adjustments that have been made so that each weapon will fire and display its muzzle flash from the tip of the weapon.

Player has been separated into layers and the body has been animated.