Enemy Path Node Traversal
Enemies now traverse their calculated path, but there are some kinks to work out, and once enemies reach their destination or touch the player, they stop and do not recalculate path. There seem to be some issues with the flocking behavior, as well. 

So far only Zumby is compatible with paths -- other enemies will need to have their update() and move() methods updated to include the calculatePath() method call and to remove the theta calculation to aim at the player.