Flow Field Heuristic Fix
Updated the heuristic used for the flow field so that it correctly, and more simply, calculates the distance to the target.

There is a bug in the flow field where if you move near a wall from a certain direction, all flow field costs overflow and become negative. This is caused by adding 1 to the cost of the previous tile when exploring, which means somehow non-traversable tiles are being added to the unexplored list.