Bug Fixes
Corrected an off-by-one error in the TMap.worldToGridCoords() method that was causing the flow field calculation to think the player was inside a wall, causing an integer overflow when calculating costs for surrounding tiles that would propagate to all tiles on the map.

Corrected flow field explore method to only look at cardinal direction neighbors and to discard diagonal neighbors, as this was throwing off the cost map and giving incorrect results.